Consolidate warn helpers into utils
